requirements.txt
asgiref==3.8.1
Django==5.2.1
django-redis==5.4.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
redis==6.2.0
sqlparse==0.5.3
